backend/requirements.txt
django
djangorestframework
django-cors-headers
python-dotenv
pymsql
mysqlclient
mysql-connector-python
Faker
